Move Avatars.kt functionality into script
